Restructure modules to centralise favicon utils
