+++ src/main.rs
@@ -1,8 +1,5 @@
 mod cli_args;
-#[cfg(feature = "server")]
-mod fallback;
 mod favicon_image;
-mod get_favicon;
 mod image_writer;
 
 #[cfg(feature = "server")]
@@ -12,7 +9,7 @@ use std::io::Write;
 
 use clap::Parser;
 use cli_args::{Cli, Command};
-use get_favicon::fetch_favicon;
+use favicon_image::FaviconImage;
 use image::ImageFormat;
 use image_writer::ImageWriter;
 
@@ -29,8 +26,9 @@ async fn main() {
             size,
             format,
         }) => {
-            // Get favicon (may be a fallback)
-            let mut favicon = match fetch_favicon(&url, size.unwrap_or(DEFAULT_IMAGE_SIZE)).await {
+            // Get favicon (will not gen a fallback)
+            let fetch_size = size.unwrap_or(DEFAULT_IMAGE_SIZE);
+            let mut favicon = match FaviconImage::fetch_for_url(&url, fetch_size).await {
                 Ok(favicon) => favicon,
                 Err(err) => {
                     eprintln!("failed to fetch favicon: {}", err);
